Updated Clean command

+++ src/commands/moderation/Clean.ts
@@ -6,7 +6,7 @@ export default class Ban extends Command {
     super({
       name: 'clean',
       description: 'Delete a specified number of messages from chat',
-      usage: ['<Limit:Number>'],
+      usage: ['<Limit:Number> <User?:ID>'],
       category: 'Moderation',
       userPermissions: ['MANAGE_MESSAGES'],
       guildOnly: true
@@ -21,11 +21,25 @@ export default class Ban extends Command {
 
     let limit = parseInt(args[0]) || 0;
 
+    if (limit < 1) return;
     if (limit > 100) limit = 100;
 
+    const target = message.guild?.members.cache.find(member => member.id === args[1]);
     const channel = message.channel as TextChannel;
-    channel.bulkDelete(limit).then(messages => {
-      message.channel.send(`Bulk deleted \`${limit}\` messages`).catch(console.error);
-    }).catch(console.error);
+    if (!target) {
+      channel.bulkDelete(limit).then(messages => {
+        message.channel.send(`Bulk deleted \`${messages.size}\` messages`).catch(console.error);
+      }).catch(console.error);
+      return;
+    }
+
+    channel.messages.fetch({ limit }).then(messages => {
+      const userMessages = messages.filter(m => m.author.id === target.id);
+      userMessages.forEach(msg => {
+        if (!msg.deletable) return;
+        msg.delete().catch(console.error);
+      });
+      message.channel.send(`Deleted \`${userMessages.size}\` messages from ${target}`);
+    });
   }
 }
